requirements.txt
appdirs==1.4.0
Django==1.10.5
djangorestframework==3.5.4
model-mommy==1.3.2
MySQL-python==1.2.5
nltk==3.2.2
numpy==1.12.0
packaging==16.8
pyparsing==2.1.10
six==1.10.0
